WIP: reorganizing static.
