+++ phyui/__init__.py
@@ -1,16 +1,20 @@
 # -*- coding: utf-8 -*-
 
-__author__ = 'Kwik Team'
-__email__ = 'cyrille.rossant at gmail.com'
+import os
+
+from IPython.html import nbextensions
+
+__author__ = 'kwikteam'
+__email__ = 'rossant at github'
 __version__ = '0.1.0-alpha'
 
 
 def prepare_js():
     """ This is needed to map js/css to the nbextensions folder
     """
-    from IPython.html import nbextensions
-    import os
-    pkgdir = os.path.join(os.path.dirname(__file__), "..", "static", "phyui")
-    nbextensions.install_nbextension(pkgdir, symlink=True, user=True, destination='phyui')
+    pkgdir = os.path.join(os.path.dirname(__file__), "static")
+    nbextensions.install_nbextension(pkgdir, symlink=True, user=True,
+                                     destination='phyui')
+
 
 prepare_js()
