+++ phyui/static/cluster/css/cluster_widget.css
@@ -1,3 +1,15 @@
+.phy-clusterview-cluster {
+    float: left;
+    width: 30px;
+    height: 30px;
+    text-align: center;
+    vertical-align: center;
+    margin: 4px;
+    padding-top: 4px;
+    border: 1px solid black;
+    cursor: pointer;
+}
+
 #cv
 {
     /* background-color: #aaa; */
