+++ phyui/static/cluster/js/widgets.js
@@ -4,8 +4,8 @@
 
 define(function(require) {
     widget = require('widgets/js/widget');
-    clusterwidget = require('/nbextensions/phyui/clusterview/d3clusterwidget.js');
-    require('/nbextensions/phyui/ipython/utils.js');
+    clusterwidget = require('/nbextensions/phyui/cluster/js/cluster_widget.js');
+    require('/nbextensions/phyui/notebook/js/utils.js');
 
     // Utility functions
     // ------------------------------------------------------------------------
