+++ phyui/static/manual_clustering.html
@@ -4,7 +4,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
-    <title>Manual Clustering</title>
+    <title>Manual clustering</title>
 
     <link rel="shortcut icon" type="image/x-icon" href="/static/base/images/favicon.ico?v=30780f272ab4aac64aa073a841546240">
 
@@ -68,10 +68,10 @@
       });
     </script>
 
-    <link rel="stylesheet" href="/nbextensions/phyui/dock-spawn/dist/css/dock-manager.css" type="text/css"/>
-    <link rel="stylesheet" href="/nbextensions/phyui/dock-spawn/dist/css/dock-wheel.css" type="text/css"/>
-    <link rel="stylesheet" href="/nbextensions/phyui/dockspawn-style.css" type="text/css"/>
-    <link rel="stylesheet" href="/nbextensions/phyui/ipython/codecell.css" type="text/css"/>
+    <link rel="stylesheet" href="/nbextensions/phyui/js/lib/dock-spawn/dist/css/dock-manager.css" type="text/css"/>
+    <link rel="stylesheet" href="/nbextensions/phyui/js/lib/dock-spawn/dist/css/dock-wheel.css" type="text/css"/>
+    <link rel="stylesheet" href="/nbextensions/phyui/dock-spawn/css/dock-spawn.css" type="text/css"/>
+    <link rel="stylesheet" href="/nbextensions/phyui/notebook/css/codecell.css" type="text/css"/>
 </head>
 
 <body style="overflow: hidden;">
@@ -114,7 +114,7 @@
   <div id="placeholder3" data-panel-caption="Code Cell" style="width: 100%; height: 50px; overflow: auto;"></div>
   <div id="placeholder4" data-panel-caption="YouNameItToo" style="width: 100%; height: 50px; overflow: auto;"></div>
   <script>
-    require(["nbextensions/phyui/manual_clustering"]);
+    require(["nbextensions/phyui/js/manual_clustering"]);
   </script>
 
 </body>
