+++ phyui/static/notebook/css/codecell.css
