+++ phyui/static/notebook/js/kernel_indicator.js
