+++ phyui/static/tests/clustertest.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <meta charset="utf-8">
 <body>
-<script src="d3.js"></script>
+<script src="../js/lib/d3.js"></script>
 <script>
   //define(function(require) {
 
@@ -17,9 +17,9 @@
 
 </script>
 
-<script src="d3clusterwidget.js"></script>
+<script src="../cluster/js/cluster_widget.js"></script>
 
-<link rel="stylesheet" href="d3clusterwidget.css"/>
+<link rel="stylesheet" href="../cluster/css/cluster_widget.css"/>
 
 <div id="cv"></div>
