--- static/phyui/clusterview/widgets.css
@@ -1,15 +1,0 @@
-/* ----------------------------------------------------------------------------
-/* Cluster view
-/* ------------------------------------------------------------------------- */
-
-.phy-clusterview-cluster {
-    float: left;
-    width: 30px;
-    height: 30px;
-    text-align: center;
-    vertical-align: center;
-    margin: 4px;
-    padding-top: 4px;
-    border: 1px solid black;
-    cursor: pointer;
-}
